Codecommited

+++ Patrika Supply Alteration/Patrika Supply Alteration/Models/SupplyViewModel.cs
@@ -3,13 +3,13 @@ namespace DCRSupplyApp.Models;
 public class SupplyViewModel
 {
     public int? BaseSupply { get; set; }
-    public decimal? SupplyMon { get; set; }
-    public decimal? SupplyTue { get; set; }
-    public decimal? SupplyWed { get; set; }
-    public decimal? SupplyThu { get; set; }
-    public decimal? SupplyFri { get; set; }
-    public decimal? SupplySat { get; set; }
-    public decimal? SupplySun { get; set; }
+    public int? SupplyMon { get; set; }
+    public int? SupplyTue { get; set; }
+    public int? SupplyWed { get; set; }
+    public int? SupplyThu { get; set; }
+    public int? SupplyFri { get; set; }
+    public int? SupplySat { get; set; }
+    public int? SupplySun { get; set; }
     public DateTime? SupplyEffectiveDate { get; set; }
     public string? SupplyFlag { get; set; }
     public string? Publ { get; set; }
